# src/controllers/album.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Put,
  Param,
  Delete,
  HttpException,
  HttpStatus,
  UseFilters,
  HttpCode,
} from '@nestjs/common';
import { HttpExceptionFilter } from '../shared/http-exception.filter';
import {
  AlbumService,
  ArtistService,
  TrackService,
  FavouritesService,
} from '../services';
import { CreateAlbumDto } from '../types/album';
import { QueryParams } from '../types/common';

@UseFilters(new HttpExceptionFilter())
@Controller('/album')
export class AlbumController {
  constructor(
    private readonly albumService: AlbumService,
    private readonly artistService: ArtistService,
    private readonly trackService: TrackService,
    private readonly favouritesService: FavouritesService,
  ) {}

  @Get()
  getAll() {
    return this.albumService.getAll();
  }

  @Get(':id')
  async getById(@Param() params: QueryParams) {
    await this.checkIfAlbumExists(params.id);

    return await this.albumService.getById(params.id);
  }

  @Post()
  async create(@Body() createAlbumDTO: CreateAlbumDto) {
    const { artistId } = createAlbumDTO;

    if (artistId === null) {
      return await this.albumService.create(createAlbumDTO);
    }

    const artist = await this.artistService.getById(artistId);

    if (!artist) {
      throw new HttpException('Artist is not found', HttpStatus.NOT_FOUND);
    }

    return await this.albumService.create(createAlbumDTO);
  }

  @Put(':id')
  async update(
    @Param() params: QueryParams,
    @Body() createUserDTO: CreateAlbumDto,
  ) {
    await this.checkIfAlbumExists(params.id);

    return await this.albumService.update(params.id, createUserDTO);
  }

  @Delete(':id')
  @HttpCode(204)
  async delete(@Param() params: QueryParams) {
    await this.checkIfAlbumExists(params.id);

    const tracks = await this.trackService.getAll();
    tracks.map((track) => {
      if (track.albumId === params.id) {
        track.albumId = null;
      }
    });

    await this.trackService.setTracks(tracks);

    const favs = await this.favouritesService.getAll();
    const favAlbums = favs.albums;

    if (favAlbums.find((id) => id === params.id)) {
      await this.favouritesService.deleteAlbum(params.id);
    }

    return await this.albumService.delete(params.id);
  }

  async checkIfAlbumExists(id: string) {
    const album = await this.albumService.getById(id);

    if (!album) {
      throw new HttpException('Album is not found', HttpStatus.NOT_FOUND);
    }
  }
}